initial commit
first steps, creating a way to iterate over the abstract syntax tree of a python program. Then will have to create an abstraction for the various node types and parse each accordingly.

Adding test_case.py as example of what the lumberjack.py file is being based on. But the code should ideally work for python code written in styles/ways different than the test_case.