fleshing out original concept still
need to find right balance of useful information to look for without being too rigid or relying on hardcoding. ideally its patterns of operations that are being summarized in the AST of each file, so key items are libraries imported, functions and end results of them (and the same for classes).